[Refactor] Refactor CCE nodepools
